+++ main/bdfd.js
@@ -6,15 +6,21 @@ const messages = {
     loadFailed: "Failed to load the BDFD Extension fully! Some features may not work. Check out the console for details."
 };
 
+const cmds = vscode.commands;
+const win = vscode.window;
+const ext = vscode.extensions;
+const work = vscode.workspace;
+
 exports.activate = activate;
 
 function activate(context) {
     bdfdFuncList(context)
     tokenColors(context)
     statusBar(context)
+    snippets(context)
 }
 
-var array = []
+var array = [];
 
 // BDFD Function List
 async function bdfdFuncList(context) {
@@ -24,22 +30,22 @@ async function bdfdFuncList(context) {
         array = await list.json();
 
         // ! Information Message
-        vscode.window.showInformationMessage(messages.funcList);
+        win.showInformationMessage(messages.funcList);
         console.debug('BDFD Function List Debug:\n', array);
 
         // ! Register
-        let disposable = vscode.commands.registerCommand('bdfd.funclist', quickPickList);
+        const disposable = cmds.registerCommand('bdfd.funclist', quickPickList);
         context.subscriptions.push(disposable);
     } catch(e) {
-        vscode.window.showErrorMessage(messages.loadFailed)
-        console.error(e)
+        win.showErrorMessage(messages.loadFailed);
+        console.error(e);
     }
 }
 
 // Quick Pick For BDFD Function List
 async function quickPickList() {
     // ! Quick Pick
-    const bdfdFunc = await vscode.window.showQuickPick(array, { matchOnDetail: true , placeHolder: "Select function"});
+    const bdfdFunc = await win.showQuickPick(array, { matchOnDetail: true , placeHolder: "Select function"});
 
     // ! API Request (Function Info)
     if (bdfdFunc == null) { return };
@@ -48,7 +54,7 @@ async function quickPickList() {
     const message = `${data.tag} > ${data.shortDescription} | Intents: ${data.intents}, Premium: ${data.premium}`;
 
     // ! Show Function Info & Open wiki action
-    const selection = await vscode.window.showInformationMessage(message, { title: "Open wiki" });
+    const selection = await win.showInformationMessage(message, { title: "Open wiki" });
     if (selection !== undefined && selection.title == "Open wiki") {
         // ! Removal of $ and [] from function tag
         const str = bdfdFunc;
@@ -57,21 +63,22 @@ async function quickPickList() {
         const name = str.substring(1, end);
 
         const wikiLink = `https://nilpointer-software.github.io/bdfd-wiki/nightly/bdscript/${name}.html`;
-        vscode.commands.executeCommand('simpleBrowser.show', wikiLink);
+        cmds.executeCommand('simpleBrowser.show', wikiLink);
     };
 }
 
 // Color Customization
 async function tokenColors(context) {
     // ! Register
-    let disposable = vscode.commands.registerCommand('bdfd.tokencolors', quickPickColors);
-    context.subscriptions.push(disposable);
+    const disposableSetColors = cmds.registerCommand('bdfd.tokencolors', quickPickColors);
+    const disposableResetColors = cmds.registerCommand('bdfd.resettokencolors', resetColors);
+    context.subscriptions.push(disposableSetColors, disposableResetColors);
 }
 
 // Quick Pick For Color Customization
 async function quickPickColors() {
     // ! Map
-    const jsonArray = vscode.extensions.getExtension('nightnutsky.bdfd-bds').packageJSON.contributes.configurationDefaults["editor.tokenColorCustomizations"].textMateRules;
+    const jsonArray = ext.getExtension('nightnutsky.bdfd-bds').packageJSON.contributes.configurationDefaults["editor.tokenColorCustomizations"].textMateRules;
     const textMateRulesMap = jsonArray.map((jsonArray) => ({
         label: jsonArray.name,
         description: `Foreground Color: ${jsonArray.settings.foreground} | Font Style: ${jsonArray.settings.fontStyle}`,
@@ -82,11 +89,11 @@ async function quickPickColors() {
     }));
 
     // ! Quick Pick
-    const color = await vscode.window.showQuickPick(textMateRulesMap, {matchOnDetail: true});
+    const color = await win.showQuickPick(textMateRulesMap, {matchOnDetail: true});
 
     // ! Modifying Color
     if (color == null) { return };
-    const modifyWhat = await vscode.window.showQuickPick([{ label: 'Foreground', description: "Modify the function's foreground color" },{ label: 'Font', description: "Modify the functions's font style" }]);
+    const modifyWhat = await win.showQuickPick([{ label: 'Foreground', description: "Modify the function's foreground color" },{ label: 'Font', description: "Modify the functions's font style" }]);
     if (modifyWhat == null) { return };
     
     const tokens = vscode.workspace.getConfiguration('editor.tokenColorCustomizations', color.target);
@@ -95,7 +102,7 @@ async function quickPickColors() {
     switch (modifyWhat.label) {
         case 'Foreground':
             // ! Input Box
-            const inputBoxForeground = await vscode.window.showInputBox({ title: 'Customizing Foreground Color', value: color.colorForeground, placeHolder: 'Example: #FFFFFF' });
+            const inputBoxForeground = await win.showInputBox({ title: 'Customizing Foreground Color', value: color.colorForeground, placeHolder: 'Example: #FFFFFF' });
 
             // ! Replace Color
             tokens.textMateRules.forEach(x => {
@@ -106,13 +113,13 @@ async function quickPickColors() {
             modified = tokens.textMateRules;
             
             // ! Modify
-            vscode.workspace.getConfiguration('editor', color.target).update('tokenColorCustomizations', {
+            work.getConfiguration('editor', color.target).update('tokenColorCustomizations', {
                 textMateRules: modified
             }, color.target);
             break;
         case 'Font':
             // ! Input Box
-            const inputBoxFontStyle = await vscode.window.showInputBox({ title: 'Customizing Font Style', value: color.colorFontStyle, placeHolder: 'Example: bold italic underline' });
+            const inputBoxFontStyle = await win.showInputBox({ title: 'Customizing Font Style', value: color.colorFontStyle, placeHolder: 'Example: bold italic underline' });
 
             // ! Replace Color
             tokens.textMateRules.forEach(x => {
@@ -123,16 +130,32 @@ async function quickPickColors() {
             modified = tokens.textMateRules;
             
             // ! Modify
-            vscode.workspace.getConfiguration('editor', color.target).update('tokenColorCustomizations', {
+            work.getConfiguration('editor', color.target).update('tokenColorCustomizations', {
                 textMateRules: modified
             }, color.target);
             break;
     };
-} 
+}
+
+// Reset Colors To Default
+async function resetColors() {
+    const reset = ext.getExtension('nightnutsky.bdfd-bds').packageJSON.contributes.configurationDefaults["editor.tokenColorCustomizations"].textMateRules;
+
+    // ! Reset
+    work.getConfiguration('editor', vscode.ConfigurationTarget.Global).update('tokenColorCustomizations', {
+        textMateRules: reset.textMateRules
+    }, vscode.ConfigurationTarget.Global);
+}
+
+// Snippets (Completion Items)
+async function snippets(context) {
+    const snippets = require('../snippets/snippets')(vscode);
+    context.subscriptions.push(snippets);
+}
 
 // Status Bar
 function statusBar(context) {
-    let bar = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Right);
+    let bar = win.createStatusBarItem(vscode.StatusBarAlignment.Right);
     bar.id = 'bdfdVersion';
     bar.name = 'BDFD Extension Version';
     bar.tooltip = 'BDFD Extension Version';